3 code

+++ GridDraw.py
@@ -72,7 +72,7 @@ class Tetris:
                 for i1 in range(i, 1, -1):
                     for j in range(self.width):
                         self.field[i1][j] = self.field[i1 - 1][j]
-        self.score += 1/2 * (lines * (lines+1))
+        self.score += int(1/2 * (lines * (lines+1)))
 
     def goSpace(self):
         if self.state.gameStarted():
